changed Metadata name to Video (to better reflect the data structure)
